logic form

+++ database/migrations/2023_01_24_060158_create_data_rectifiers_table.php
@@ -16,9 +16,9 @@ return new class extends Migration
         Schema::create('data_rectifiers', function (Blueprint $table) {
             $table->id();
             $table->foreignId('rectifier_id');
-            $table->float('voltage');
-            $table->float('current');
-            $table->float('temp'); 
+            $table->double('voltage');
+            $table->double('current');
+            $table->double('temp');
             $table->timestamps();
         });
     }
